Added more services and removed additional include directories.-Aaron
Introduced Factory, EventManager, SceneManager to ServiceLocator. Trying to remove additional includes for libraries but VS filters giving me trouble with glm.

+++ Shared_Engine/Systems/Graphics/FontRendering/FontTest.cpp
@@ -1,8 +1,8 @@
 #include "FontTest.h"
 #include "../../Graphics_Header.h"
 #include "../ShaderProgram/ShaderProgram.h"
-#include "freetype2/ft2build.h"
-#include "freetype2/freetype/freetype.h"
+#include "../freetype2/ft2build.h"
+#include "freetype/freetype.h"
 
 #include <iostream>
 #include <map>
@@ -72,7 +72,7 @@ void LoadFonts()
 			texture,
 			glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
 			glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-			face->glyph->advance.x
+			(GLuint)face->glyph->advance.x
 		};
 		Characters.insert(std::pair<GLchar, Character>(c, character));
 	}
